+++ src/type_checker.cpp
@@ -438,12 +438,13 @@ CallableTypes TypeChecker::resolveCallee(ApplyNode &node) {
 
   // otherwise, resolve function type
   CallableTypes callableTypes(this->typePool.getUnresolvedType());
-  auto &type = this->checkType(this->typePool.get(TYPE::Func), exprNode);
+  //  auto &type = this->checkType(this->typePool.get(TYPE::Func), exprNode);
+  auto &type = this->checkTypeExactly(exprNode);
   if (type.isFuncType()) {
     node.setKind(ApplyNode::FUNC_CALL);
     callableTypes = cast<FunctionType>(type).toCallableTypes();
   } else {
-    this->reportError<NotCallable>(exprNode);
+    this->reportError<NotCallable>(exprNode, type.getName());
   }
   return callableTypes;
 }